Settings buttons

+++ scripts/testvtt.js
@@ -17,12 +17,27 @@ class ToDoList{
     }
     static initialize() {
         this.toDoListConfig = new ToDoListConfig();
+        game.settings.register(this.ID, this.SETTINGS.INJECT_BUTTON, {
+            name: `TODO-LIST.settings.${this.SETTINGS.INJECT_BUTTON}.Name`,
+            default: true,
+            type: Boolean,
+            scope: 'client',
+            config: true,
+            hint: `TODO-LIST.settings.${this.SETTINGS.INJECT_BUTTON}.Hint`,
+            onChange: () => ui.players.render()
+          });
+    }
+    static SETTINGS = {
+        INJECT_BUTTON: 'inject-button'
     }
 }
 Hooks.once('init', () => {
     ToDoList.initialize();
 });
 Hooks.on('renderPlayerList', (playerList, html) => {
+    if (!game.settings.get(ToDoList.ID, ToDoList.SETTINGS.INJECT_BUTTON)) {
+        return;
+      }
     // find the element which has our logged in user's id
   const loggedInUserListItem = html.find(`[data-user-id="${game.userId}"]`)
   // create localized tooltip
@@ -149,9 +164,17 @@ class ToDoListConfig extends FormApplication {
             }
       
             case 'delete': {
-              await ToDoListData.deleteToDo(toDoId);
-              this.render();
-              break;
+                const confirmed = await Dialog.confirm({
+                  title: game.i18n.localize("TODO-LIST.confirms.deleteConfirm.Title"),
+                  content: game.i18n.localize("TODO-LIST.confirms.deleteConfirm.Content")
+                });
+        
+                if (confirmed) {
+                  await ToDoListData.deleteToDo(toDoId);
+                  this.render();
+                }
+        
+                break;
             }
       
             default:
